# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family Health Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, suspended]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [users]
      summary: Suspend a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve a pending user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/activate:
    post:
      operationId: activateUser
      tags: [users]
      summary: Reactivate a suspended user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weight:
    get:
      operationId: listWeightEntries
      tags: [weight]
      summary: List weight entries for the current user
      parameters:
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
          description: Family member user ID (if viewing own family member)
      responses:
        "200":
          description: List of weight entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWeightEntry
      tags: [weight]
      summary: Add a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeightEntryBody"
      responses:
        "201":
          description: Weight entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weight/{id}:
    patch:
      operationId: updateWeightEntry
      tags: [weight]
      summary: Update a weight entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWeightEntryBody"
      responses:
        "200":
          description: Updated weight entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWeightEntry
      tags: [weight]
      summary: Delete a weight entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, user]
        gender:
          type: ["string", "null"]
          enum: [male, female, null]
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - role
        - gender
        - status
        - createdAt
...
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, user]
        gender:
          type: string
          enum: [male, female]
        status:
          type: string
          enum: [pending, active, suspended]
      required:
        - username
        - password
        - displayName
        - role
        - gender
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        role:
          type: string
          enum: [admin, user]
        gender:
          type: string
          enum: [male, female]
        status:
          type: string
          enum: [pending, active, suspended]
    ResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
    WeightEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        weightKg:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - weightKg
        - recordedAt
        - note
        - createdAt
    CreateWeightEntryBody:
      type: object
      properties:
        weightKg:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        memberId:
          type: ["integer", "null"]
      required:
        - weightKg
        - recordedAt
    UpdateWeightEntryBody:
      type: object
      properties:
        weightKg:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
...